Removed form-data; fixed buggy scripts to send data in the Strava-appropriate manner for POST, GET, and DELETE requests.

+++ scripts/deleteStravaWebhook.js
@@ -1,29 +1,28 @@
 import {} from 'dotenv/config.js'
 import axios from 'axios'
-import FormData from 'form-data'
 
-// Script to create the webhook callback from the Strava Webhook API.
-// Should only run in production environments!
+// Script to delete the webhook callback from the Strava Webhook API.
 // See: https://developers.strava.com/docs/webhooks/
 
 const deleteStravaWebhook = () => {
+  if (process.argv.length !== 3 || process.argv[2] == null) {
+    console.log('Usage: $ node scripts/deleteStravaWebhook.js <subscriptionId>')
+    return -1
+  }
+
   const clientId = process.env.STRAVA_CLIENT_ID
   const clientSecret = process.env.STRAVA_CLIENT_SECRET
+  const subscriptionId = process.argv[2]
 
-  // TODO: Fill in the subscription ID here.
-  const subscriptionId = '12345'
-
-  // Params have to be sent as a form for some reason
-  // Trying to use the curl -F option
-  const form = new FormData()
-  form.append('client_id', clientId)
-  form.append('client_secret', clientSecret)
+  console.log(`Attempting to delete Strava webhook subscription with id: "${subscriptionId}"`)
 
   axios({
     method: 'delete',
     url: `https://www.strava.com/api/v3/push_subscriptions/${subscriptionId}`,
-    data: form,
-    headers: {'Content-Type': 'multipart/form-data' },
+    data: {
+      client_id: clientId,
+      client_secret: clientSecret,
+    },
   }).then((response) => {
     // Response will happen after callback URL echoes the Strava request
     // sent to it. See callbackURL route for details.
